Allow noise to be 0
Bayesian Optimisation in presence of noiseless samples is a thing.
Consider for example deterministic computer simulations. For these, it
makes no sense to ever resample the same location and it gives the model
full confidence in the sample locations. This allows the model to fit to
data that it normally struggles with and either gives up or gives the
whole sample a huge variance making it basically pick at random.

My previous work-around for this strictly-positive assert was to set the
noise to `f64::EPSILON` which did work somewhat but it made variance
near samples `NaN`.

I have verified in my program that setting the noise to 0 keeps the
optimisation working and that variance at samples is now 0 which is
exactly what I wanted.