Add optional serde support for kernel/prior
There are many possible things that one may want to do with this but a
very straight-forward one is to write out the kernel/prior we're using
and coming back to the GP in the future, reading back the values.

Another use-case is allowing a user to specify kernel parameters
dynamically from some config file when starting an optimisation and it's
easy if it's `Deserialize` and we're using `serde` already.